requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0.post1
python-multipart==0.0.9
starlette==0.36.3
Pillow==10.2.0
numpy==1.26.4
openai>=1.0.0
